requirements.txt
streamlit
pandasai
python-dotenv
openai==0.28